WIP: Opprydding etter nyflyt lansering
- Refaktorer ut de forskjellige helper klassene for å lette på den kognitive lasten
