+++ src/test/kotlin/no/nav/lydia/helper/TestContainerHelper.kt
@@ -1,95 +1,23 @@
 package no.nav.lydia.helper
 
 import com.github.kittinunf.fuel.core.Request
-import com.github.kittinunf.fuel.core.ResponseResultOf
-import com.github.kittinunf.fuel.core.extensions.authentication
-import com.github.kittinunf.fuel.core.extensions.jsonBody
 import com.github.kittinunf.fuel.httpDelete
 import com.github.kittinunf.fuel.httpGet
 import com.github.kittinunf.fuel.httpPost
 import com.github.kittinunf.fuel.httpPut
 import com.github.kittinunf.fuel.serialization.responseObject
 import ia.felles.definisjoner.bransjer.Bransje
-import io.kotest.matchers.collections.shouldHaveSize
-import io.kotest.matchers.shouldBe
 import io.kotest.matchers.string.shouldContain
 import io.kotest.matchers.string.shouldNotContain
-import kotlinx.datetime.LocalDate
-import kotlinx.datetime.LocalDateTime
-import kotlinx.datetime.toKotlinLocalDateTime
 import kotlinx.serialization.InternalSerializationApi
 import kotlinx.serialization.builtins.ListSerializer
 import kotlinx.serialization.json.Json
 import kotlinx.serialization.serializer
 import no.nav.lydia.Topic
-import no.nav.lydia.container.ia.sak.kartlegging.BehovsvurderingApiTest.Companion.ID_TIL_SPØRSMÅL_MED_FLERVALG_MULIGHETER
-import no.nav.lydia.helper.TestContainerHelper.Companion.applikasjon
-import no.nav.lydia.helper.TestContainerHelper.Companion.authContainerHelper
-import no.nav.lydia.helper.TestContainerHelper.Companion.kafkaContainerHelper
-import no.nav.lydia.helper.TestContainerHelper.Companion.performDelete
-import no.nav.lydia.helper.TestContainerHelper.Companion.performGet
-import no.nav.lydia.helper.TestContainerHelper.Companion.performPost
-import no.nav.lydia.helper.TestContainerHelper.Companion.performPut
-import no.nav.lydia.helper.TestContainerHelper.Companion.postgresContainerHelper
 import no.nav.lydia.helper.TestData.Companion.lagPerioder
-import no.nav.lydia.ia.sak.api.IASakDto
-import no.nav.lydia.ia.sak.api.IASakLeveranserPerTjenesteDto
-import no.nav.lydia.ia.sak.api.IA_SAK_LEVERANSE_PATH
-import no.nav.lydia.ia.sak.api.IA_SAK_RADGIVER_PATH
-import no.nav.lydia.ia.sak.api.KanGjennomføreStatusendring
-import no.nav.lydia.ia.sak.api.SAMARBEIDSHISTORIKK_PATH
-import no.nav.lydia.ia.sak.api.SaksStatusDto
-import no.nav.lydia.ia.sak.api.SakshistorikkDto
-import no.nav.lydia.ia.sak.api.dokument.DOKUMENT_PUBLISERING_BASE_ROUTE
-import no.nav.lydia.ia.sak.api.dokument.DokumentPubliseringDto
-import no.nav.lydia.ia.sak.api.ny.flyt.NY_FLYT_API_PATH
-import no.nav.lydia.ia.sak.api.ny.flyt.NY_FLYT_PATH
-import no.nav.lydia.ia.sak.api.plan.EndreTemaRequest
-import no.nav.lydia.ia.sak.api.plan.EndreUndertemaRequest
-import no.nav.lydia.ia.sak.api.plan.PLAN_BASE_ROUTE
-import no.nav.lydia.ia.sak.api.plan.PlanDto
-import no.nav.lydia.ia.sak.api.plan.PlanDtoI
-import no.nav.lydia.ia.sak.api.plan.PlanMedPubliseringStatusDto
-import no.nav.lydia.ia.sak.api.plan.PlanUndertemaDto
-import no.nav.lydia.ia.sak.api.samarbeid.IASamarbeidDto
-import no.nav.lydia.ia.sak.api.spørreundersøkelse.OppdaterBehovsvurderingDto
-import no.nav.lydia.ia.sak.api.spørreundersøkelse.SPØRREUNDERSØKELSE_BASE_ROUTE
-import no.nav.lydia.ia.sak.api.spørreundersøkelse.SpørreundersøkelseDto
-import no.nav.lydia.ia.sak.api.spørreundersøkelse.SpørreundersøkelseResultatDto
-import no.nav.lydia.ia.sak.api.spørreundersøkelse.SpørreundersøkelseSvarDto
-import no.nav.lydia.ia.sak.api.spørreundersøkelse.SpørreundersøkelseUtenInnholdDto
-import no.nav.lydia.ia.sak.domene.plan.InnholdMalDto
-import no.nav.lydia.ia.sak.domene.plan.PlanMalDto
-import no.nav.lydia.ia.sak.domene.plan.PlanUndertema
-import no.nav.lydia.ia.sak.domene.plan.TemaMalDto
-import no.nav.lydia.ia.sak.domene.spørreundersøkelse.Spørreundersøkelse
-import no.nav.lydia.ia.team.BrukerITeamDto
-import no.nav.lydia.ia.team.IA_SAK_TEAM_PATH
-import no.nav.lydia.ia.årsak.domene.ValgtÅrsak
-import no.nav.lydia.integrasjoner.kartlegging.HendelsType
-import no.nav.lydia.integrasjoner.kartlegging.SpørreundersøkelseHendeleseNøkkel
-import no.nav.lydia.integrasjoner.kvittering.KvitteringDto
 import no.nav.lydia.integrasjoner.ssb.NæringsDownloader
 import no.nav.lydia.integrasjoner.ssb.NæringsRepository
-import no.nav.lydia.statusoversikt.StatusoversiktResponsDto
-import no.nav.lydia.statusoversikt.api.STATUSOVERSIKT_PATH
 import no.nav.lydia.sykefraværsstatistikk.LANDKODE_NO
-import no.nav.lydia.sykefraværsstatistikk.Publiseringsinfo
-import no.nav.lydia.sykefraværsstatistikk.api.ANTALL_TREFF
-import no.nav.lydia.sykefraværsstatistikk.api.FILTERVERDIER_PATH
-import no.nav.lydia.sykefraværsstatistikk.api.FilterverdierDto
-import no.nav.lydia.sykefraværsstatistikk.api.HISTORISK_STATISTIKK
-import no.nav.lydia.sykefraværsstatistikk.api.PUBLISERINGSINFO
-import no.nav.lydia.sykefraværsstatistikk.api.SISTE_4_KVARTALER
-import no.nav.lydia.sykefraværsstatistikk.api.SISTE_TILGJENGELIGE_KVARTAL
-import no.nav.lydia.sykefraværsstatistikk.api.SYKEFRAVÆRSSTATISTIKK_PATH
-import no.nav.lydia.sykefraværsstatistikk.api.Søkeparametere
-import no.nav.lydia.sykefraværsstatistikk.api.Søkeparametere.Companion.VIRKSOMHETER_PER_SIDE
-import no.nav.lydia.sykefraværsstatistikk.api.VirksomhetsoversiktDto
-import no.nav.lydia.sykefraværsstatistikk.api.VirksomhetsoversiktResponsDto
-import no.nav.lydia.sykefraværsstatistikk.api.VirksomhetsstatistikkSiste4KvartalDto
-import no.nav.lydia.sykefraværsstatistikk.domene.HistoriskStatistikk
-import no.nav.lydia.sykefraværsstatistikk.domene.VirksomhetsstatistikkSisteKvartal
 import no.nav.lydia.sykefraværsstatistikk.import.Kategori
 import no.nav.lydia.virksomhet.domene.Sektor
 import org.slf4j.Logger
@@ -99,11 +27,7 @@ import org.testcontainers.containers.Network
 import org.testcontainers.containers.output.Slf4jLogConsumer
 import org.testcontainers.containers.wait.strategy.HttpWaitStrategy
 import org.testcontainers.images.builder.ImageFromDockerfile
-import java.time.ZoneId
-import java.time.ZonedDateTime
-import java.util.UUID
 import kotlin.io.path.Path
-import kotlin.test.fail
 
 class TestContainerHelper {
     companion object {
@@ -255,1207 +179,6 @@ class TestContainerHelper {
         infix fun GenericContainer<*>.shouldContainLog(regex: Regex) = logs shouldContain regex
 
         infix fun GenericContainer<*>.shouldNotContainLog(regex: Regex) = logs shouldNotContain regex
-    }
-}
-
-class SakHelper {
-    companion object {
-        fun bliMedITeam(
-            token: String,
-            saksnummer: String,
-        ) = applikasjon.performPost("$IA_SAK_TEAM_PATH/$saksnummer")
-            .authentication().bearer(token)
-            .tilSingelRespons<BrukerITeamDto>().third.fold(
-                success = { respons -> respons },
-                failure = { fail(it.message) },
-            )
-            .also { it.saksnummer shouldBe saksnummer }
-
-        fun IASakDto.bliEierResponse(token: String) =
-            applikasjon.performPost("$NY_FLYT_PATH/$orgnr/bli-eier")
-                .authentication().bearer(token = token)
-                .tilSingelRespons<IASakDto>()
-
-        fun IASakDto.bliEier(token: String) = bliEierResponse(token).third.get()
-
-        fun IASakDto.leggTilFolger(token: String) = also { bliMedITeam(token = token, saksnummer) }
-
-        fun IASakDto.hentSaksStatus(token: String = authContainerHelper.saksbehandler1.token) =
-            applikasjon
-                .performGet("$IA_SAK_RADGIVER_PATH/$orgnr/$saksnummer/status")
-                .authentication().bearer(token = token)
-                .tilSingelRespons<SaksStatusDto>().third
-                .fold(
-                    success = { it },
-                    failure = {
-                        fail(it.response.body().asString("text/plain"))
-                    },
-                )
-
-        fun hentIASakLeveranser(
-            orgnr: String,
-            saksnummer: String,
-            token: String = authContainerHelper.saksbehandler1.token,
-        ) = applikasjon.performGet("$IA_SAK_RADGIVER_PATH/$IA_SAK_LEVERANSE_PATH/$orgnr/$saksnummer")
-            .authentication().bearer(token = token)
-            .tilListeRespons<IASakLeveranserPerTjenesteDto>().third.fold(
-                success = { respons -> respons },
-                failure = {
-                    fail(it.stackTraceToString())
-                },
-            )
-
-        fun hentSak(
-            orgnummer: String,
-            saksnummer: String? = null,
-            token: String = authContainerHelper.saksbehandler1.token,
-        ): IASakDto {
-            val triple = applikasjon.performGet("$NY_FLYT_PATH/virksomhet/$orgnummer/samarbeidsperiode${saksnummer?.let { "/$it" } ?: ""}")
-                .authentication().bearer(token = token)
-                .responseObject(IASakDto.serializer())
-
-            if (triple.statuskode() == 200) {
-                return triple.third.get()
-            } else if (triple.statuskode() == 204) {
-                fail("Ingen aktive saker funnet")
-            } else {
-                fail(triple.third.toString())
-            }
-        }
-
-        fun hentSamarbeidshistorikkNyFlyt(
-            orgnummer: String,
-            token: String = authContainerHelper.saksbehandler1.token,
-        ) = hentSamarbeidshistorikkNyFlytRespons(orgnummer = orgnummer, token = token).third.fold(
-            success = { respons -> respons },
-            failure = { fail(it.stackTraceToString()) },
-        )
-
-        fun hentSamarbeidshistorikkNyFlytRespons(
-            orgnummer: String,
-            token: String = authContainerHelper.saksbehandler1.token,
-        ): ResponseResultOf<List<SakshistorikkDto>> {
-            val url = "$NY_FLYT_PATH/virksomhet/$orgnummer/historikk"
-            return applikasjon.performGet(url)
-                .authentication().bearer(token = token)
-                .tilListeRespons<SakshistorikkDto>()
-        }
-
-        fun hentSamarbeidshistorikk(
-            orgnummer: String,
-            token: String = authContainerHelper.saksbehandler1.token,
-        ) = hentSamarbeidshistorikkRespons(orgnummer, token).third.fold(
-            success = { respons -> respons },
-            failure = { fail(it.stackTraceToString()) },
-        )
-
-        fun hentSamarbeidshistorikkRespons(
-            orgnummer: String,
-            token: String = authContainerHelper.saksbehandler1.token,
-        ) = applikasjon.performGet("$IA_SAK_RADGIVER_PATH/$SAMARBEIDSHISTORIKK_PATH/$orgnummer")
-            .authentication().bearer(token = token)
-            .tilListeRespons<SakshistorikkDto>()
-
-        fun hentSamarbeidshistorikkForOrgnrRespons(
-            orgnr: String,
-            token: String = authContainerHelper.saksbehandler1.token,
-        ) = applikasjon.performGet("$IA_SAK_RADGIVER_PATH/$SAMARBEIDSHISTORIKK_PATH/$orgnr")
-            .authentication().bearer(token = token)
-            .tilListeRespons<SakshistorikkDto>()
-
-        fun IASakDto.kanGjennomføreStatusendring(
-            samarbeidDto: IASamarbeidDto,
-            statusEndring: String,
-            token: String = authContainerHelper.saksbehandler1.token,
-        ) = applikasjon.performGet("$IA_SAK_RADGIVER_PATH/${this.orgnr}/${this.saksnummer}/${samarbeidDto.id}/kan/$statusEndring")
-            .authentication().bearer(token)
-            .tilSingelRespons<KanGjennomføreStatusendring>()
-            .third.fold(
-                success = { respons -> respons },
-                failure = {
-                    fail(it.stackTraceToString())
-                },
-            )
-
-        fun IASakDto.oppdaterHendelsesTidspunkter(
-            antallDagerTilbake: Long,
-            token: String = authContainerHelper.saksbehandler1.token,
-        ): IASakDto {
-            postgresContainerHelper.performUpdate(
-                """
-                update ia_sak_hendelse 
-                    set opprettet=(opprettet - interval '$antallDagerTilbake' day)
-                    where saksnummer='${this.saksnummer}';
-                """.trimIndent(),
-            )
-            postgresContainerHelper.performUpdate(
-                """
-                update ia_sak 
-                    set endret=(opprettet - interval '$antallDagerTilbake' day)
-                    where saksnummer='${this.saksnummer}';
-                """.trimIndent(),
-            )
-            return requireNotNull(hentSak(this.orgnr, saksnummer = this.saksnummer, token = token))
-        }
-    }
-}
-
-class DokumentPubliseringHelper {
-    companion object {
-        fun publiserDokument(
-            dokumentReferanseId: String,
-            dokumentType: DokumentPubliseringDto.Type = DokumentPubliseringDto.Type.BEHOVSVURDERING,
-            token: String,
-        ) = applikasjon.performPost(url = "$DOKUMENT_PUBLISERING_BASE_ROUTE/type/$dokumentType/ref/$dokumentReferanseId")
-            .authentication().bearer(token = token)
-            .tilSingelRespons<DokumentPubliseringDto>()
-
-        fun sendKvittering(
-            dokument: DokumentPubliseringDto,
-            samarbeidId: Int,
-            ønsketPublisertDato: LocalDateTime =
-                ZonedDateTime.now().withZoneSameInstant(ZoneId.of("Europe/Oslo")).toLocalDateTime().toKotlinLocalDateTime(),
-        ) {
-            val dokId = UUID.randomUUID().toString()
-            kafkaContainerHelper.sendOgVentTilKonsumert(
-                nøkkel = dokId,
-                melding = Json.encodeToString(
-                    KvitteringDto(
-                        referanseId = dokument.referanseId,
-                        samarbeidId = samarbeidId,
-                        dokumentId = dokId,
-                        journalpostId = UUID.randomUUID().toString(),
-                        publisertDato = ønsketPublisertDato,
-                        type = dokument.dokumentType.name,
-                    ),
-                ),
-                topic = Topic.DOKUMENT_KVITTERING_TOPIC,
-            )
-        }
-    }
-}
-
-class IASakSpørreundersøkelseHelper {
-    companion object {
-        fun opprettSpørreundersøkelse(
-            orgnr: String,
-            saksnummer: String,
-            samarbeidId: Int,
-            token: String = authContainerHelper.saksbehandler1.token,
-            type: Spørreundersøkelse.Type,
-        ) = applikasjon.performPost(
-            "$NY_FLYT_API_PATH/virksomhet/$orgnr/samarbeidsperiode/$saksnummer/samarbeid/$samarbeidId/kartlegging/${type.name}",
-        ).authentication().bearer(token)
-
-        fun hentSpørreundersøkelser(
-            orgnr: String,
-            saksnummer: String,
-            prosessId: Int,
-            token: String = authContainerHelper.saksbehandler1.token,
-        ) = applikasjon.performGet("$SPØRREUNDERSØKELSE_BASE_ROUTE/$orgnr/$saksnummer/prosess/$prosessId/")
-            .authentication().bearer(token)
-            .tilListeRespons<SpørreundersøkelseUtenInnholdDto>().third.fold(
-                success = { it },
-                failure = { fail(it.message) },
-            )
-
-        fun SpørreundersøkelseDto.hentKartleggingresultatPdf(
-            orgnr: String,
-            saksnummer: String,
-            token: String = authContainerHelper.saksbehandler1.token,
-        ) = applikasjon.performGet("$SPØRREUNDERSØKELSE_BASE_ROUTE/$orgnr/$saksnummer/$id/pdf")
-            .authentication().bearer(token)
-            .response().third.fold(
-                success = { it },
-                failure = { fail(it.message) },
-            )
-
-        fun hentSpørreundersøkelse(
-            orgnr: String,
-            saksnummer: String,
-            prosessId: Int,
-            token: String = authContainerHelper.saksbehandler1.token,
-            type: Spørreundersøkelse.Type,
-        ) = applikasjon.performGet("$SPØRREUNDERSØKELSE_BASE_ROUTE/$orgnr/$saksnummer/prosess/$prosessId/type/${type.name}")
-            .authentication().bearer(token)
-            .tilListeRespons<SpørreundersøkelseUtenInnholdDto>().third.fold(
-                success = { it },
-                failure = { fail(it.message) },
-            )
-
-        fun IASamarbeidDto.opprettKartlegging(
-            orgnr: String,
-            type: Spørreundersøkelse.Type,
-            token: String = authContainerHelper.saksbehandler1.token,
-        ) = opprettSpørreundersøkelse(
-            orgnr = orgnr,
-            saksnummer = saksnummer,
-            samarbeidId = id,
-            token = token,
-            type = type,
-        ).tilSingelRespons<SpørreundersøkelseDto>().third.get()
-
-        fun IASakDto.opprettKartlegging(
-            samarbeidId: Int = hentAlleSamarbeid().first().id,
-            type: Spørreundersøkelse.Type,
-            token: String = authContainerHelper.saksbehandler1.token,
-        ) = opprettSpørreundersøkelse(
-            orgnr = orgnr,
-            saksnummer = saksnummer,
-            samarbeidId = samarbeidId,
-            token = token,
-            type = type,
-        ).tilSingelRespons<SpørreundersøkelseDto>().third.fold(
-            success = { respons -> respons },
-            failure = { fail(it.message) },
-        )
-
-        fun IASakDto.opprettBehovsvurdering(
-            samarbeidId: Int = hentAlleSamarbeid().first().id,
-            token: String = authContainerHelper.saksbehandler1.token,
-        ) = opprettSpørreundersøkelse(
-            orgnr = orgnr,
-            saksnummer = saksnummer,
-            samarbeidId = samarbeidId,
-            token = token,
-            type = Spørreundersøkelse.Type.Behovsvurdering,
-        ).tilSingelRespons<SpørreundersøkelseDto>().third.fold(
-            success = { respons -> respons },
-            failure = { fail(it.message) },
-        )
-
-        fun IASakDto.opprettSvarOgAvsluttSpørreundersøkelse(
-            type: Spørreundersøkelse.Type,
-            token: String = authContainerHelper.saksbehandler1.token,
-            samarbeidId: Int = hentAlleSamarbeid().first().id,
-            temaIdx: Int = 0,
-            spørsmålIdx: Int = 0,
-            svaralternativIdx: Int = 0,
-            antallSvarPåSpørsmål: Int = 3,
-        ): SpørreundersøkelseDto =
-            when (type) {
-                Spørreundersøkelse.Type.Evaluering -> opprettEvaluering(
-                    prosessId = samarbeidId,
-                    token = token,
-                )
-
-                Spørreundersøkelse.Type.Behovsvurdering -> opprettBehovsvurdering(
-                    samarbeidId = samarbeidId,
-                    token = token,
-                )
-            }
-                .start(
-                    token = token,
-                    orgnummer = orgnr,
-                    saksnummer = saksnummer,
-                )
-                .also {
-                    it.svarPåSpørsmål(
-                        temaIdx = temaIdx,
-                        spørsmålIdx = spørsmålIdx,
-                        svaralternativIdx = svaralternativIdx,
-                        antallSvarPåSpørsmål = antallSvarPåSpørsmål,
-                    )
-                }
-                .fullfør(
-                    token = token,
-                    orgnummer = orgnr,
-                    saksnummer = saksnummer,
-                )
-
-        fun IASakDto.opprettEvaluering(
-            prosessId: Int = hentAlleSamarbeid().first().id,
-            token: String = authContainerHelper.saksbehandler1.token,
-        ) = opprettSpørreundersøkelse(
-            orgnr = orgnr,
-            saksnummer = saksnummer,
-            samarbeidId = prosessId,
-            token = token,
-            type = Spørreundersøkelse.Type.Evaluering,
-        ).tilSingelRespons<SpørreundersøkelseDto>().third.fold(
-            success = { respons -> respons },
-            failure = { fail(it.message) },
-        )
-
-        fun IASakDto.hentForhåndsvisning(
-            prosessId: Int = hentAlleSamarbeid().first().id,
-            type: Spørreundersøkelse.Type,
-            spørreundersøkseId: String,
-            token: String = authContainerHelper.saksbehandler1.token,
-        ): SpørreundersøkelseDto =
-            applikasjon.performGet("$SPØRREUNDERSØKELSE_BASE_ROUTE/$orgnr/$saksnummer/prosess/$prosessId/type/${type.name}/$spørreundersøkseId")
-                .authentication().bearer(token)
-                .tilSingelRespons<SpørreundersøkelseDto>().third.fold(
-                    success = { it },
-                    failure = { fail(it.message) },
-                )
-
-        fun SpørreundersøkelseDto.start(
-            token: String = authContainerHelper.saksbehandler1.token,
-            orgnummer: String,
-            saksnummer: String,
-        ): SpørreundersøkelseDto =
-            applikasjon.performPost("$NY_FLYT_API_PATH/virksomhet/$orgnummer/samarbeidsperiode/$saksnummer/samarbeid/$samarbeidId/kartlegging/$id/start")
-                .authentication().bearer(token)
-                .tilSingelRespons<SpørreundersøkelseDto>().third.fold(
-                    success = { it },
-                    failure = { fail(it.message) },
-                )
-
-        fun SpørreundersøkelseDto.fullfør(
-            token: String = authContainerHelper.saksbehandler1.token,
-            orgnummer: String,
-            saksnummer: String,
-        ) = applikasjon.performPost("$NY_FLYT_API_PATH/virksomhet/$orgnummer/samarbeidsperiode/$saksnummer/samarbeid/$samarbeidId/kartlegging/$id/fullfor")
-            .authentication().bearer(token)
-            .tilSingelRespons<SpørreundersøkelseDto>().third.fold(
-                success = { it },
-                failure = { fail(it.message) },
-            )
-
-        fun SpørreundersøkelseDto.svarPåSpørsmål(
-            temaIdx: Int = 0,
-            spørsmålIdx: Int = 0,
-            svaralternativIdx: Int = 0,
-            antallSvarPåSpørsmål: Int = 3,
-        ) {
-            val førsteSpørsmål = this.temaer[temaIdx].spørsmålOgSvaralternativer[spørsmålIdx]
-            val førsteSvaralternativ = førsteSpørsmål.svaralternativer[svaralternativIdx]
-
-            repeat(antallSvarPåSpørsmål) {
-                val sesjonId = UUID.randomUUID()
-                sendKartleggingSvarTilKafka(
-                    kartleggingId = this.id,
-                    spørsmålId = førsteSpørsmål.id,
-                    sesjonId = sesjonId.toString(),
-                    svarIder = listOf(førsteSvaralternativ.svarId),
-                )
-            }
-        }
-
-        fun SpørreundersøkelseDto.slett(
-            token: String = authContainerHelper.saksbehandler1.token,
-            orgnummer: String,
-            saksnummer: String,
-        ) = slettResponse(token = token, orgnummer = orgnummer, saksnummer = saksnummer).third.get()
-
-        fun SpørreundersøkelseDto.slettResponse(
-            token: String = authContainerHelper.saksbehandler1.token,
-            orgnummer: String,
-            saksnummer: String,
-        ) = applikasjon.performDelete("$NY_FLYT_API_PATH/virksomhet/$orgnummer/samarbeidsperiode/$saksnummer/samarbeid/$samarbeidId/kartlegging/$id")
-            .authentication().bearer(token)
-            .tilSingelRespons<SpørreundersøkelseDto>()
-
-        fun SpørreundersøkelseDto.flytt(
-            token: String = authContainerHelper.saksbehandler1.token,
-            orgnummer: String,
-            saksnummer: String,
-            samarbeidId: Int,
-        ) = applikasjon.performPut("$SPØRREUNDERSØKELSE_BASE_ROUTE/$id")
-            .authentication().bearer(token)
-            .jsonBody(
-                Json.encodeToString(
-                    OppdaterBehovsvurderingDto(
-                        orgnummer = orgnummer,
-                        saksnummer = saksnummer,
-                        prosessId = samarbeidId,
-                    ),
-                ),
-            ).tilSingelRespons<SpørreundersøkelseDto>().third.fold(
-                success = { it },
-                failure = { fail("${it.message}: ${it.response.body().asString("text/plain")}") },
-            )
-
-        fun SpørreundersøkelseDto.svarAlternativerTilEtFlervalgSpørsmål(): List<String> =
-            this.svarAlternativerTilEtSpørsmål(ID_TIL_SPØRSMÅL_MED_FLERVALG_MULIGHETER).map { it.svarId }
-
-        fun SpørreundersøkelseDto.svarAlternativerTilEtSpørsmål(spørsmålId: String) =
-            this.temaer.map { tema ->
-                tema.spørsmålOgSvaralternativer.firstOrNull { it.id == spørsmålId }
-            }.first()!!.svaralternativer
-
-        fun SpørreundersøkelseDto.sendKartleggingFlervalgSvarTilKafka(
-            sesjonId: String = UUID.randomUUID().toString(),
-            svarIder: List<String> = this.svarAlternativerTilEtFlervalgSpørsmål(),
-        ) = sendKartleggingSvarTilKafka(
-            kartleggingId = id,
-            spørsmålId = ID_TIL_SPØRSMÅL_MED_FLERVALG_MULIGHETER,
-            sesjonId = sesjonId,
-            svarIder = svarIder,
-        )
-
-        fun SpørreundersøkelseDto.sendKartleggingSvarTilKafka(
-            spørsmålId: String = temaer.first().spørsmålOgSvaralternativer.first().id,
-            sesjonId: String = UUID.randomUUID().toString(),
-            svarIder: List<String> =
-                listOf(temaer.first().spørsmålOgSvaralternativer.first().svaralternativer.first().svarId),
-        ) = sendKartleggingSvarTilKafka(
-            kartleggingId = id,
-            spørsmålId = spørsmålId,
-            sesjonId = sesjonId,
-            svarIder = svarIder,
-        )
-
-        fun SpørreundersøkelseDto.stengTema(temaId: Int) {
-            temaer.filter { it.temaId == temaId } shouldHaveSize 1
-            kafkaContainerHelper.sendOgVentTilKonsumert(
-                nøkkel = Json.encodeToString(
-                    SpørreundersøkelseHendeleseNøkkel(
-                        this.id,
-                        HendelsType.STENG_TEMA,
-                    ),
-                ),
-                melding = Json.encodeToString(temaId),
-                topic = Topic.SPORREUNDERSOKELSE_HENDELSE_TOPIC,
-            )
-        }
-
-        fun sendKartleggingSvarTilKafka(
-            kartleggingId: String = UUID.randomUUID().toString(),
-            spørsmålId: String = UUID.randomUUID().toString(),
-            sesjonId: String = UUID.randomUUID().toString(),
-            svarIder: List<String> = listOf(UUID.randomUUID().toString(), UUID.randomUUID().toString()),
-        ): SpørreundersøkelseSvarDto {
-            val spørreundersøkelseSvarDto = SpørreundersøkelseSvarDto(
-                spørreundersøkelseId = kartleggingId,
-                spørsmålId = spørsmålId,
-                sesjonId = sesjonId,
-                svarIder = svarIder,
-            )
-            kafkaContainerHelper.sendOgVentTilKonsumert(
-                nøkkel = "${sesjonId}_$spørsmålId",
-                melding = Json.encodeToString(
-                    spørreundersøkelseSvarDto,
-                ),
-                topic = Topic.SPORREUNDERSOKELSE_SVAR_TOPIC,
-            )
-            return spørreundersøkelseSvarDto
-        }
-
-        fun hentSpørreundersøkelseResultat(
-            token: String = authContainerHelper.saksbehandler1.token,
-            orgnr: String,
-            saksnummer: String,
-            spørreundersøkelseId: String,
-        ): SpørreundersøkelseResultatDto =
-            applikasjon.performGet("$SPØRREUNDERSØKELSE_BASE_ROUTE/$orgnr/$saksnummer/$spørreundersøkelseId")
-                .authentication().bearer(token)
-                .tilSingelRespons<SpørreundersøkelseResultatDto>().third.fold(
-                    success = { respons -> respons },
-                    failure = { fail(it.message) },
-                )
-    }
-}
-
-class PlanHelper {
-    companion object {
-        val START_DATO = LocalDate(year = 2021, monthNumber = 1, dayOfMonth = 1)
-        val SLUTT_DATO = LocalDate(year = 2022, monthNumber = 2, dayOfMonth = 2)
-
-        fun PlanDtoI.antallTemaInkludert() = temaer.filter { it.inkludert }.size
-
-        fun PlanDtoI.antallInnholdInkludert() = temaer.flatMap { it.undertemaer }.filter { it.inkludert }.size
-
-        fun PlanDtoI.antallInnholdMedStatus(status: PlanUndertema.Status) =
-            temaer.flatMap { it.undertemaer }.filter {
-                it.inkludert &&
-                    it.status == status
-            }.size
-
-        fun PlanDtoI.tidligstStartDato(): LocalDate =
-            this.temaer.flatMap { it.undertemaer }
-                .filter { it.startDato != null }
-                .minOf { it.startDato!! }
-
-        fun PlanDtoI.senesteSluttDato(): LocalDate =
-            this.temaer.flatMap { it.undertemaer }
-                .filter { it.sluttDato != null }
-                .maxOf { it.sluttDato!! }
-
-        fun IASakDto.hentPlan(
-            prosessId: Int = hentAlleSamarbeid().first().id,
-            token: String = authContainerHelper.saksbehandler1.token,
-        ) = hentPlanResponse(
-            samarbeidId = prosessId,
-            token = token,
-        ).third.fold(
-            success = { respons -> respons },
-            failure = { fail(it.message) },
-        )
-
-        fun IASakDto.hentPlanResponse(
-            samarbeidId: Int = hentAlleSamarbeid().first().id,
-            token: String = authContainerHelper.saksbehandler1.token,
-        ) = applikasjon.performGet("$PLAN_BASE_ROUTE/$orgnr/$saksnummer/prosess/$samarbeidId")
-            .authentication().bearer(token)
-            .tilSingelRespons<PlanMedPubliseringStatusDto>()
-
-        private fun hentPlan(
-            orgnr: String,
-            saksnummer: String,
-            prosessId: Int,
-            token: String = authContainerHelper.saksbehandler1.token,
-        ) = applikasjon.performGet("$PLAN_BASE_ROUTE/$orgnr/$saksnummer/prosess/$prosessId")
-            .authentication().bearer(token)
-            .tilSingelRespons<PlanMedPubliseringStatusDto>().third.fold(
-                success = { respons -> respons },
-                failure = { fail(it.message) },
-            )
-
-        fun IASakDto.endreStatusPåInnholdIPlan(
-            token: String = authContainerHelper.saksbehandler1.token,
-            temaId: Int,
-            innholdId: Int,
-            status: PlanUndertema.Status,
-        ) = endreStatus(
-            token = token,
-            orgnr = orgnr,
-            saksnummer = saksnummer,
-            temaId = temaId,
-            undertemaId = innholdId,
-            status = status,
-            prosessId = hentAlleSamarbeid().first().id,
-        )
-
-        fun PlanMalDto.inkluderAlt(): PlanMalDto =
-            this.copy(
-                tema = this.tema.map { tema ->
-                    TemaMalDto(
-                        rekkefølge = tema.rekkefølge,
-                        navn = tema.navn,
-                        inkludert = true,
-                        innhold = tema.innhold.map { innhold ->
-                            InnholdMalDto(
-                                rekkefølge = innhold.rekkefølge,
-                                navn = innhold.navn,
-                                inkludert = true,
-                                startDato = LocalDate(2021, 1, 1),
-                                sluttDato = LocalDate(2021, 1, 2),
-                            )
-                        },
-                    )
-                },
-            )
-
-        fun PlanMalDto.inkluderEttTemaOgEttInnhold(
-            temanummer: Int,
-            innholdnummer: Int,
-            startDato: LocalDate = START_DATO,
-            sluttDato: LocalDate = SLUTT_DATO,
-        ): PlanMalDto =
-            this.copy(
-                tema = tema.map { tema ->
-                    if (temanummer == tema.rekkefølge) {
-                        tema.copy(
-                            inkludert = true,
-                            innhold = tema.innhold.inkluderEttInnhold(
-                                innholdnummer,
-                                startDato = startDato,
-                                sluttDato = sluttDato,
-                            ),
-                        )
-                    } else {
-                        tema
-                    }
-                },
-            )
-
-        fun PlanMalDto.inkluderEttTemaOgAltInnhold(
-            temanummer: Int,
-            startDato: LocalDate = START_DATO,
-            sluttDato: LocalDate = SLUTT_DATO,
-        ): PlanMalDto =
-            this.copy(
-                tema = tema.map { tema ->
-                    if (temanummer == tema.rekkefølge) {
-                        tema.copy(
-                            inkludert = true,
-                            innhold = tema.innhold.inkluderAltInnholdIMal(startDato = startDato, sluttDato = sluttDato),
-                        )
-                    } else {
-                        tema
-                    }
-                },
-            )
-
-        private fun List<InnholdMalDto>.inkluderAltInnholdIMal(
-            startDato: LocalDate = START_DATO,
-            sluttDato: LocalDate = SLUTT_DATO,
-        ): List<InnholdMalDto> =
-            this.map { innhold ->
-                innhold.copy(
-                    inkludert = true,
-                    startDato = startDato,
-                    sluttDato = sluttDato,
-                )
-            }
-
-        private fun List<InnholdMalDto>.inkluderEttInnhold(
-            innholdnummer: Int,
-            startDato: LocalDate = START_DATO,
-            sluttDato: LocalDate = SLUTT_DATO,
-        ): List<InnholdMalDto> =
-            this.map { innhold ->
-                if (innholdnummer == innhold.rekkefølge) {
-                    innhold.copy(
-                        inkludert = true,
-                        startDato = startDato,
-                        sluttDato = sluttDato,
-                    )
-                } else {
-                    innhold
-                }
-            }
-
-        fun List<PlanUndertemaDto>.inkluderAltInnhold(
-            startDato: LocalDate = START_DATO,
-            sluttDato: LocalDate = SLUTT_DATO,
-        ): List<PlanUndertemaDto> =
-            this.map { innhold ->
-                innhold.copy(
-                    inkludert = true,
-                    startDato = startDato,
-                    sluttDato = sluttDato,
-                )
-            }
-
-        fun PlanMedPubliseringStatusDto.inkluderTemaOgAltInnhold(
-            temaId: Int,
-            startDato: LocalDate = START_DATO,
-            sluttDato: LocalDate = SLUTT_DATO,
-        ): PlanMedPubliseringStatusDto =
-            this.copy(
-                temaer = temaer.map { tema ->
-                    if (tema.id == temaId) {
-                        tema.copy(
-                            inkludert = true,
-                            undertemaer = tema.undertemaer.inkluderAltInnhold(
-                                startDato = startDato,
-                                sluttDato = sluttDato,
-                            ),
-                        )
-                    } else {
-                        tema
-                    }
-                },
-            )
-
-        fun PlanDto.inkluderAlt(
-            startDato: LocalDate = START_DATO,
-            sluttDato: LocalDate = SLUTT_DATO,
-        ): PlanDto =
-            this.copy(
-                temaer = temaer.map { tema ->
-                    tema.copy(
-                        inkludert = true,
-                        undertemaer = tema.undertemaer.inkluderAltInnhold(
-                            startDato = startDato,
-                            sluttDato = sluttDato,
-                        ),
-                    )
-                },
-            )
-
-        fun PlanMedPubliseringStatusDto.inkluderAlt(
-            startDato: LocalDate = START_DATO,
-            sluttDato: LocalDate = SLUTT_DATO,
-        ): PlanMedPubliseringStatusDto =
-            this.copy(
-                temaer = temaer.map { tema ->
-                    tema.copy(
-                        inkludert = true,
-                        undertemaer = tema.undertemaer.inkluderAltInnhold(
-                            startDato = startDato,
-                            sluttDato = sluttDato,
-                        ),
-                    )
-                },
-            )
-
-        // TODO: [OPPRYDDING] Bør gå bort ifra denne måten å opprette planer
-        fun IASakDto.opprettEnPlan(
-            token: String = authContainerHelper.saksbehandler1.token,
-            plan: PlanMalDto = hentPlanMal().inkluderAlt(),
-        ) = hentAlleSamarbeid().first().opprettSamarbeidsplan(
-            orgnr = orgnr,
-            planMal = plan,
-            token = token,
-        )
-
-        fun IASamarbeidDto.opprettSamarbeidsplan(
-            orgnr: String,
-            planMal: PlanMalDto = hentPlanMal().inkluderAlt(),
-            token: String = authContainerHelper.saksbehandler1.token,
-        ) = applikasjon.performPost("$NY_FLYT_API_PATH/virksomhet/$orgnr/samarbeidsperiode/${this.saksnummer}/samarbeid/${this.id}/plan")
-            .authentication().bearer(token)
-            .jsonBody(
-                Json.encodeToString(planMal),
-            ).tilSingelRespons<PlanMedPubliseringStatusDto>().third.fold(
-                success = { it },
-                failure = { fail(it.message) },
-            )
-
-        fun IASakDto.endreEttTemaIPlan(
-            temaId: Int,
-            endring: List<EndreUndertemaRequest>,
-            prosessId: Int = hentAlleSamarbeid().first().id,
-            token: String = authContainerHelper.saksbehandler1.token,
-        ) = applikasjon.performPut("$PLAN_BASE_ROUTE/$orgnr/$saksnummer/prosess/$prosessId/$temaId")
-            .jsonBody(Json.encodeToString(endring))
-            .authentication().bearer(token)
-            .tilSingelRespons<PlanDto>().third.fold(
-                success = { respons -> respons },
-                failure = { fail(it.message) },
-            )
-
-        fun IASakDto.endreFlereTemaerIPlan(
-            endring: List<EndreTemaRequest>,
-            prosessId: Int = hentAlleSamarbeid().first().id,
-            token: String = authContainerHelper.saksbehandler1.token,
-        ) = applikasjon.performPut("$PLAN_BASE_ROUTE/$orgnr/$saksnummer/prosess/$prosessId")
-            .jsonBody(Json.encodeToString(endring))
-            .authentication().bearer(token)
-            .tilSingelRespons<PlanDto>().third.fold(
-                success = { respons -> respons },
-                failure = { fail(it.message) },
-            )
-
-        private fun endrePlan(
-            orgnr: String,
-            saksnummer: String,
-            prosessId: Int,
-            endring: List<EndreTemaRequest>,
-            token: String = authContainerHelper.saksbehandler1.token,
-        ) = applikasjon.performPut("$PLAN_BASE_ROUTE/$orgnr/$saksnummer/prosess/$prosessId")
-            .jsonBody(Json.encodeToString(endring))
-            .authentication().bearer(token)
-            .tilSingelRespons<PlanDto>().third.fold(
-                success = { respons -> respons },
-                failure = { fail(it.message) },
-            )
-
-        private fun endreStatus(
-            orgnr: String,
-            saksnummer: String,
-            prosessId: Int,
-            status: PlanUndertema.Status,
-            temaId: Int,
-            undertemaId: Int,
-            token: String = authContainerHelper.saksbehandler1.token,
-        ) = applikasjon.performPut("$PLAN_BASE_ROUTE/$orgnr/$saksnummer/prosess/$prosessId/$temaId/$undertemaId")
-            .jsonBody(Json.encodeToString(status))
-            .authentication().bearer(token)
-            .tilSingelRespons<PlanDto>().third.fold(
-                success = { respons -> respons },
-                failure = { fail(it.message) },
-            )
-
-        fun PlanDtoI.tilRequest(): List<EndreTemaRequest> =
-            this.temaer.map { tema ->
-                EndreTemaRequest(
-                    tema.id,
-                    tema.inkludert,
-                    tema.undertemaer.tilRequest(),
-                )
-            }
-
-        fun List<PlanUndertemaDto>.tilRequest(): List<EndreUndertemaRequest> =
-            this.map { innhold ->
-                EndreUndertemaRequest(
-                    innhold.id,
-                    innhold.inkludert,
-                    innhold.startDato,
-                    innhold.sluttDato,
-                )
-            }
-
-        fun hentPlanMal(token: String = authContainerHelper.saksbehandler1.token) =
-            applikasjon.performGet("$PLAN_BASE_ROUTE/mal")
-                .authentication().bearer(token)
-                .tilSingelRespons<PlanMalDto>().third.fold(
-                    success = { respons -> respons },
-                    failure = { fail(it.message) },
-                )
-
-        fun PlanMedPubliseringStatusDto.planleggOgFullførAlleUndertemaer(
-            orgnummer: String,
-            saksnummer: String,
-            prosessId: Int,
-            token: String = authContainerHelper.saksbehandler1.token,
-        ) = this.copy(
-            temaer = temaer.map { tema ->
-                tema.copy(
-                    inkludert = true,
-                    undertemaer = tema.undertemaer.map {
-                        it.copy(
-                            inkludert = true,
-                            startDato = START_DATO,
-                            sluttDato = SLUTT_DATO,
-                        )
-                    },
-                )
-            },
-        ).tilRequest().also { endringer ->
-            endrePlan(
-                orgnr = orgnummer,
-                saksnummer = saksnummer,
-                prosessId = prosessId,
-                endring = endringer,
-                token = token,
-            )
-        }.forEach { tema ->
-            tema.undertemaer.forEach {
-                endreStatus(
-                    orgnr = orgnummer,
-                    saksnummer = saksnummer,
-                    prosessId = prosessId,
-                    status = PlanUndertema.Status.FULLFØRT,
-                    temaId = tema.id,
-                    undertemaId = it.id,
-                )
-            }
-        }.let {
-            hentPlan(
-                orgnr = orgnummer,
-                saksnummer = saksnummer,
-                prosessId = prosessId,
-                token = token,
-            )
-        }
-    }
-}
-
-class StatusoversiktHelper {
-    companion object {
-        fun hentStatusoversikt(
-            kommuner: String = "",
-            fylker: String = "",
-            næringsgrupper: String = "",
-            sykefraværsprosentFra: String = "",
-            sykefraværsprosentTil: String = "",
-            ansatteFra: String = "",
-            ansatteTil: String = "",
-            bransjeProgram: String = "",
-            snittFilter: String = "",
-            eiere: String = "",
-            sektor: String = "",
-            token: String = authContainerHelper.saksbehandler1.token,
-        ) = hentStatusoversiktRespons(
-            kommuner = kommuner,
-            fylker = fylker,
-            næringsgrupper = næringsgrupper,
-            sykefraværsprosentFra = sykefraværsprosentFra,
-            sykefraværsprosentTil = sykefraværsprosentTil,
-            ansatteFra = ansatteFra,
-            ansatteTil = ansatteTil,
-            bransjeProgram = bransjeProgram,
-            snittFilter = snittFilter,
-            eiere = eiere,
-            sektor = sektor,
-            token = token,
-        )
-
-        private fun hentStatusoversiktRespons(
-            kommuner: String = "",
-            fylker: String = "",
-            næringsgrupper: String = "",
-            sykefraværsprosentFra: String = "",
-            sykefraværsprosentTil: String = "",
-            ansatteFra: String = "",
-            ansatteTil: String = "",
-            bransjeProgram: String = "",
-            snittFilter: String = "",
-            eiere: String = "",
-            sektor: String = "",
-            token: String = authContainerHelper.saksbehandler1.token,
-        ) = applikasjon.performGet(
-            STATUSOVERSIKT_PATH +
-                "?${Søkeparametere.KOMMUNER}=$kommuner" +
-                "&${Søkeparametere.FYLKER}=$fylker" +
-                "&${Søkeparametere.NÆRINGSGRUPPER}=$næringsgrupper" +
-                "&${Søkeparametere.SYKEFRAVÆRSPROSENT_FRA}=$sykefraværsprosentFra" +
-                "&${Søkeparametere.SYKEFRAVÆRSPROSENT_TIL}=$sykefraværsprosentTil" +
-                "&${Søkeparametere.ANSATTE_FRA}=$ansatteFra" +
-                "&${Søkeparametere.ANSATTE_TIL}=$ansatteTil" +
-                "&${Søkeparametere.BRANSJEPROGRAM}=$bransjeProgram" +
-                "&${Søkeparametere.SNITT_FILTER}=$snittFilter" +
-                "&${Søkeparametere.IA_SAK_EIERE}=$eiere" +
-                "&${Søkeparametere.SEKTOR}=$sektor",
-        )
-            .authentication().bearer(token)
-            .tilSingelRespons<StatusoversiktResponsDto>()
-    }
-}
-
-class StatistikkHelper {
-    companion object {
-        fun hentSykefravær(
-            success: (VirksomhetsoversiktResponsDto) -> Unit,
-            kommuner: String = "",
-            fylker: String = "",
-            næringsgrupper: String = "",
-            sorteringsnokkel: String = "",
-            sorteringsretning: String = "",
-            sykefraværsprosentFra: String = "",
-            sykefraværsprosentTil: String = "",
-            ansatteFra: String = "",
-            ansatteTil: String = "",
-            iaStatus: String = "",
-            side: String = "",
-            bransjeProgram: String = "",
-            eiere: String = "",
-            token: String = authContainerHelper.saksbehandler1.token,
-        ) = hentSykefraværRespons(
-            kommuner = kommuner,
-            fylker = fylker,
-            næringsgrupper = næringsgrupper,
-            sorteringsnokkel = sorteringsnokkel,
-            sorteringsretning = sorteringsretning,
-            sykefraværsprosentFra = sykefraværsprosentFra,
-            sykefraværsprosentTil = sykefraværsprosentTil,
-            ansatteFra = ansatteFra,
-            ansatteTil = ansatteTil,
-            iaStatus = iaStatus,
-            side = side,
-            bransjeProgram = bransjeProgram,
-            eiere = eiere,
-            token = token,
-        ).third
-            .fold(success = { response -> success.invoke(response) }, failure = {
-                fail("${it.message} - ${it.response.body().asString("text/plain")}")
-            })
-
-        fun hentSykefravær(
-            kommuner: String = "",
-            fylker: String = "",
-            næringsgrupper: String = "",
-            sorteringsnokkel: String = "",
-            sorteringsretning: String = "",
-            sykefraværsprosentFra: String = "",
-            sykefraværsprosentTil: String = "",
-            snittFilter: String = "",
-            ansatteFra: String = "",
-            ansatteTil: String = "",
-            iaStatus: String = "",
-            virksomhetTilstand: String = "",
-            side: String = "",
-            bransjeProgram: String = "",
-            eiere: String = "",
-            sektor: List<Sektor> = listOf(),
-            token: String = authContainerHelper.saksbehandler1.token,
-        ) = hentSykefraværRespons(
-            kommuner = kommuner,
-            fylker = fylker,
-            næringsgrupper = næringsgrupper,
-            sorteringsnokkel = sorteringsnokkel,
-            sorteringsretning = sorteringsretning,
-            sykefraværsprosentFra = sykefraværsprosentFra,
-            sykefraværsprosentTil = sykefraværsprosentTil,
-            snittFilter = snittFilter,
-            ansatteFra = ansatteFra,
-            ansatteTil = ansatteTil,
-            iaStatus = iaStatus,
-            virkomhetTilstand = virksomhetTilstand,
-            side = side,
-            bransjeProgram = bransjeProgram,
-            eiere = eiere,
-            sektor = sektor,
-            token = token,
-        ).third.get()
-
-        fun hentSykefraværRespons(
-            kommuner: String = "",
-            fylker: String = "",
-            næringsgrupper: String = "",
-            sorteringsnokkel: String = "",
-            sorteringsretning: String = "",
-            sykefraværsprosentFra: String = "",
-            sykefraværsprosentTil: String = "",
-            snittFilter: String = "",
-            ansatteFra: String = "",
-            ansatteTil: String = "",
-            iaStatus: String = "",
-            virkomhetTilstand: String = "",
-            side: String = "",
-            bransjeProgram: String = "",
-            eiere: String = "",
-            sektor: List<Sektor> = listOf(),
-            token: String = authContainerHelper.saksbehandler1.token,
-        ) = applikasjon.performGet(
-            SYKEFRAVÆRSSTATISTIKK_PATH +
-                "?${Søkeparametere.KOMMUNER}=$kommuner" +
-                "&${Søkeparametere.FYLKER}=$fylker" +
-                "&${Søkeparametere.NÆRINGSGRUPPER}=$næringsgrupper" +
-                "&${Søkeparametere.SORTERINGSNØKKEL}=$sorteringsnokkel" +
-                "&${Søkeparametere.SORTERINGSRETNING}=$sorteringsretning" +
-                "&${Søkeparametere.SYKEFRAVÆRSPROSENT_FRA}=$sykefraværsprosentFra" +
-                "&${Søkeparametere.SYKEFRAVÆRSPROSENT_TIL}=$sykefraværsprosentTil" +
-                "&${Søkeparametere.SNITT_FILTER}=$snittFilter" +
-                "&${Søkeparametere.ANSATTE_FRA}=$ansatteFra" +
-                "&${Søkeparametere.ANSATTE_TIL}=$ansatteTil" +
-                "&${Søkeparametere.IA_STATUS}=$iaStatus" +
-                "&${Søkeparametere.VIRKSOMHET_TILSTAND}=$virkomhetTilstand" +
-                "&${Søkeparametere.SIDE}=$side" +
-                "&${Søkeparametere.BRANSJEPROGRAM}=$bransjeProgram" +
-                "&${Søkeparametere.IA_SAK_EIERE}=$eiere" +
-                "&${Søkeparametere.SEKTOR}=${sektor.joinToString(separator = ",") { it.kode }}",
-        )
-            .authentication().bearer(token)
-            .tilSingelRespons<VirksomhetsoversiktResponsDto>()
-
-        fun hentStatistikkHistorikk(orgnr: String) = hentStatistikkHistorikkRespons(orgnr).third.get()
-
-        private fun hentStatistikkHistorikkRespons(
-            orgnr: String,
-            token: String = authContainerHelper.saksbehandler1.token,
-        ) = applikasjon.performGet(
-            "$SYKEFRAVÆRSSTATISTIKK_PATH/$orgnr/$HISTORISK_STATISTIKK",
-        )
-            .authentication().bearer(token)
-            .tilSingelRespons<HistoriskStatistikk>()
-
-        fun hentSykefraværForVirksomhetSiste4KvartalerRespons(
-            orgnummer: String,
-            token: String = authContainerHelper.saksbehandler1.token,
-        ) = applikasjon.performGet("$SYKEFRAVÆRSSTATISTIKK_PATH/$orgnummer/$SISTE_4_KVARTALER")
-            .authentication().bearer(token)
-            .tilSingelRespons<VirksomhetsstatistikkSiste4KvartalDto>()
-
-        private fun hentPubliseringsinfoRespons(token: String = authContainerHelper.saksbehandler1.token) =
-            applikasjon.performGet("$SYKEFRAVÆRSSTATISTIKK_PATH/$PUBLISERINGSINFO")
-                .authentication().bearer(token)
-                .tilSingelRespons<Publiseringsinfo>()
-
-        private fun hentSykefraværForVirksomhetSisteTilgjengeligKvartalRespons(
-            orgnummer: String,
-            token: String = authContainerHelper.saksbehandler1.token,
-        ) = applikasjon.performGet("$SYKEFRAVÆRSSTATISTIKK_PATH/$orgnummer/$SISTE_TILGJENGELIGE_KVARTAL")
-            .authentication().bearer(token)
-            .tilSingelRespons<VirksomhetsstatistikkSisteKvartal>()
-
-        fun hentSykefraværForVirksomhetSiste4Kvartaler(
-            orgnummer: String,
-            token: String = authContainerHelper.saksbehandler1.token,
-        ) = hentSykefraværForVirksomhetSiste4KvartalerRespons(orgnummer = orgnummer, token = token).third
-            .fold(success = { response -> response }, failure = { fail(it.message) })
-
-        fun hentPubliseringsinfo(token: String = authContainerHelper.saksbehandler1.token) =
-            hentPubliseringsinfoRespons(token).third
-                .fold(success = { response -> response }, failure = { fail(it.message) })
-
-        fun hentSykefraværForVirksomhetSisteTilgjengeligKvartal(
-            orgnummer: String,
-            token: String = authContainerHelper.saksbehandler1.token,
-        ) = hentSykefraværForVirksomhetSisteTilgjengeligKvartalRespons(orgnummer = orgnummer, token = token).third
-            .fold(success = { response -> response }, failure = { fail(it.message) })
-
-        fun hentSykefraværForAlleVirksomheter(): List<VirksomhetsoversiktDto> {
-            var side = 1
-            val liste = mutableListOf<VirksomhetsoversiktDto>()
-
-            do {
-                val sykefravær = hentSykefravær(side = "${side++}")
-                liste.addAll(sykefravær.data)
-            } while (sykefravær.data.size == VIRKSOMHETER_PER_SIDE)
-
-            return liste.toList()
-        }
-
-        // Defaultverdiane her er standard-verdiane i kall frå frontend per 2023-11-09
-        fun hentSykefraværForAlleVirksomheterMedFilter(
-            side: Int = 1,
-            kommuner: String = "",
-            fylker: String = "",
-            næringsgrupper: String = "",
-            sorteringsnokkel: String = "",
-            sorteringsretning: String = "",
-            sykefraværsprosentFra: String = "0.00",
-            sykefraværsprosentTil: String = "100.00",
-            snittFilter: String = "",
-            ansatteFra: String = "5",
-            ansatteTil: String = "",
-            iaStatus: String = "",
-            bransjeProgram: String = "",
-            eiere: String = "",
-            sektor: List<Sektor> = listOf(),
-            token: String = authContainerHelper.saksbehandler1.token,
-        ): List<VirksomhetsoversiktDto> {
-            var itererbarSide = side
-            val liste = mutableListOf<VirksomhetsoversiktDto>()
-
-            do {
-                val sykefravær = hentSykefravær(
-                    side = "${itererbarSide++}",
-                    kommuner = kommuner,
-                    fylker = fylker,
-                    næringsgrupper = næringsgrupper,
-                    sorteringsnokkel = sorteringsnokkel,
-                    sorteringsretning = sorteringsretning,
-                    sykefraværsprosentFra = sykefraværsprosentFra,
-                    sykefraværsprosentTil = sykefraværsprosentTil,
-                    snittFilter = snittFilter,
-                    ansatteFra = ansatteFra,
-                    ansatteTil = ansatteTil,
-                    iaStatus = iaStatus,
-                    bransjeProgram = bransjeProgram,
-                    eiere = eiere,
-                    sektor = sektor,
-                    token = token,
-                )
-                liste.addAll(sykefravær.data)
-            } while (sykefravær.data.size == VIRKSOMHETER_PER_SIDE)
-
-            return liste.toList()
-        }
-
-        fun hentTotaltAntallTreffISykefravær(
-            kommuner: String = "",
-            fylker: String = "",
-            næringsgrupper: String = "",
-            sorteringsnokkel: String = "",
-            sorteringsretning: String = "",
-            sykefraværsprosentFra: String = "",
-            sykefraværsprosentTil: String = "",
-            ansatteFra: String = "",
-            ansatteTil: String = "",
-            iaStatus: String = "",
-            virksomhetTilstand: String = "",
-            side: String = "",
-            bransjeProgram: String = "",
-            eiere: String = "",
-            sektor: List<Sektor> = listOf(),
-            token: String = authContainerHelper.saksbehandler1.token,
-        ): Int =
-            applikasjon.performGet(
-                "$SYKEFRAVÆRSSTATISTIKK_PATH/$ANTALL_TREFF" +
-                    "?${Søkeparametere.KOMMUNER}=$kommuner" +
-                    "&${Søkeparametere.FYLKER}=$fylker" +
-                    "&${Søkeparametere.NÆRINGSGRUPPER}=$næringsgrupper" +
-                    "&${Søkeparametere.SORTERINGSNØKKEL}=$sorteringsnokkel" +
-                    "&${Søkeparametere.SORTERINGSRETNING}=$sorteringsretning" +
-                    "&${Søkeparametere.SYKEFRAVÆRSPROSENT_FRA}=$sykefraværsprosentFra" +
-                    "&${Søkeparametere.SYKEFRAVÆRSPROSENT_TIL}=$sykefraværsprosentTil" +
-                    "&${Søkeparametere.ANSATTE_FRA}=$ansatteFra" +
-                    "&${Søkeparametere.ANSATTE_TIL}=$ansatteTil" +
-                    "&${Søkeparametere.IA_STATUS}=$iaStatus" +
-                    "&${Søkeparametere.VIRKSOMHET_TILSTAND}=$virksomhetTilstand" +
-                    "&${Søkeparametere.SIDE}=$side" +
-                    "&${Søkeparametere.BRANSJEPROGRAM}=$bransjeProgram" +
-                    "&${Søkeparametere.IA_SAK_EIERE}=$eiere" +
-                    "&${Søkeparametere.SEKTOR}=${sektor.joinToString(separator = ",") { it.kode }}",
-            )
-                .authentication().bearer(token)
-                .tilSingelRespons<Int>()
-                .third
-                .fold(success = { response -> response }, failure = { fail(it.message) })
-
-        fun hentFilterverdier(token: String = authContainerHelper.saksbehandler1.token) =
-            applikasjon.performGet("$SYKEFRAVÆRSSTATISTIKK_PATH/$FILTERVERDIER_PATH")
-                .authentication().bearer(token)
-                .tilSingelRespons<FilterverdierDto>()
-                .third
-                .fold(success = { response -> response }, failure = { fail(it.message) })
     }
 }
